Split kafka cluster health into two new checks
