+++ Kafka/legos/kafka_broker_health_check/README.md
@@ -1,14 +1,13 @@
 [<img align="left" src="https://unskript.com/assets/favicon.png" width="100" height="100" style="padding-right: 5px">]
 (https://unskript.com/assets/favicon.png)
-<h1>Kafka get cluster health</h1>
+<h1>Get Kafka broker health</h1>
 
 ## Description
-Fetches the health of the Kafka cluster including brokers, topics, and partitions.
+Checks the health of the Kafka brokers by determining if the Kafka producer can establish a connection with the bootstrap brokers of a Kafka cluster.
 
 ## Lego Details
-	kafka_get_cluster_health(handle: KafkaProducer)
+	kafka_broker_health_check(handle)
 		handle: Object of type unSkript KAFKA Connector.
-
 
 ## Lego Input
 This Lego takes inputs handle.
