--- Kafka/legos/kafka_get_cluster_health/kafka_get_cluster_health.json
@@ -1,12 +1,0 @@
-{
-  "action_title": "Kafka get cluster health",
-  "action_description": "Fetches the health of the Kafka cluster including brokers, topics, and partitions.",
-  "action_type": "LEGO_TYPE_KAFKA",
-  "action_entry_function": "kafka_get_cluster_health",
-  "action_needs_credential": true,
-  "action_output_type": "ACTION_OUTPUT_TYPE_DICT",
-  "action_is_check": false,
-  "action_supports_iteration": true,
-  "action_supports_poll": true,
-  "action_categories":[ "CATEGORY_TYPE_DEVOPS", "CATEGORY_TYPE_SRE","CATEGORY_TYPE_KAFKA"]
-}
