+++ Kafka/legos/kafka_topic_partition_health_check/README.md
@@ -1,0 +1,22 @@
+[<img align="left" src="https://unskript.com/assets/favicon.png" width="100" height="100" style="padding-right: 5px">]
+(https://unskript.com/assets/favicon.png)
+<h1>Get Kafka topic partition health</h1>
+
+## Description
+Checks the health of the Kafka topics and their partitions.This check checks if the topics have any partitions at all.
+
+## Lego Details
+	kafka_topic_partition_health_check(handle)
+		handle: Object of type unSkript KAFKA Connector.
+
+
+## Lego Input
+This Lego takes inputs handle.
+
+## Lego Output
+Here is a sample output.
+<img src="./1.png">
+
+## See it in Action
+
+You can see this Lego in action following this link [unSkript Live](https://us.app.unskript.io)
